photos/app/requirements.txt
fastapi
uvicorn
gunicorn
pydantic
osxphotos
Pillow  # Keep regular Pillow for Docker compatibility
pillow-heif
opencv-python-headless  # Headless version for Docker/server environments
apscheduler
markdown
numpy  # Required for OpenCV
